Fixed input json "parameters" handling
If the "parameters" key is not present in the input json file, jsonResponse.parameters is equal to undefined, not null. Abstract equality comparison should be used instead of strict.
Joining of parameters tokens should be done in all cases, even if "parameters" key is not present or if it's null. Otherwise the tokens are joined with "," which breaks the program logic.